+++ core/src/main/java/org/seedstack/seed/spring/api/WithApplicationContexts.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright (c) 2013-2015 by The SeedStack authors. All rights reserved.
+ *
+ * This file is part of SeedStack, An enterprise-oriented full development stack.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.seedstack.seed.spring.api;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Specifies the Spring applications contexts to use in an integration test. Apply this annotation on a test class. It
+ * can also be applied to methods if a different kernel is started for each method (i.e. testing kernel mode is per
+ * test, not per test class).
+ *
+ * @author adrien.lauer@mpsa.com
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ ElementType.TYPE, ElementType.METHOD })
+@Documented
+@Inherited
+public @interface WithApplicationContexts {
+    /**
+     * @return the resource path of the application contexts to use.
+     */
+    String[] value();
+}