+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 2.1.0 (2015-11-26)
 
+* [new] `@WithApplicationContexts` is now supported on test methods in conjunction with `@WithCommandLine` enabling a per-method alteration of the test Spring context.
 * [chg] Refactored as an add-on and updated to work with Seed 2.1.0+
 
 # Version 2.0.0 (2015-07-30)